requirements.txt
azure_storage==0.37.0
msrest==0.6.21
termcolor==1.1.0
